+++ tasks/behat.js
@@ -1,5 +1,4 @@
 module.exports = function(grunt) {
-
   /**
    * Define "behat" tasks.
    *
@@ -18,22 +17,32 @@ module.exports = function(grunt) {
    *       "subsite": "http://sub.dev-site.local"
    *     },
    *     "behat": {
+   *       "flags": "--tags ~@wip",
+   *       "path": "/usr/local/bin/behat",
    *       "subsite": {
    *          "src": "./features/subsite/*.feature",
    *          "debug": false
    *       }
    *     }
    *   }
+   *
+   * If the command line option `--behat_flags="flags"` is set, the `flags`
+   * specified will be passed to behat and override any flags from the config.
    */
   grunt.loadNpmTasks('grunt-parallel-behat');
-  var config = grunt.config.get('config'),
-    flags = '',
-    _ = require('lodash');
+  var config = grunt.config.get('config');
+  var _ = require('lodash');
 
   if (config.buildPaths.html && config.siteUrls) {
     for (var key in config.siteUrls) {
       if (config.siteUrls.hasOwnProperty(key)) {
         var options = {};
+        var path = 'vendor/bin/behat';
+
+        // Allow configuring the path to the behat binary.
+        if (config.behat && config.behat.path) {
+          path = config.behat.path;
+        }
 
         // Check for per-site behat options.
         if (config.behat && config.behat[key]) {
@@ -57,10 +66,10 @@ module.exports = function(grunt) {
             options: _.extend({
               config: './test/behat.yml',
               maxProcesses: 5,
-              bin: 'vendor/bin/behat',
+              bin: path,
               debug: true,
               env: _.extend({}, process.env, {
-                "BEHAT_PARAMS": "{\"extensions\": {\"Drupal\\\\DrupalExtension\": {\"drupal\": {\"drupal_root\": \"./" + config.buildPaths.html + "\"}}, \"Behat\\\\MinkExtension\": {\"base_url\": \"" + config.siteUrls[key] + "\", \"zombie\": {\"node_modules_path\": \"" + process.cwd() + "/node_modules/\"}}}}"
+                BEHAT_PARAMS: "{\"extensions\": {\"Drupal\\\\DrupalExtension\": {\"drupal\": {\"drupal_root\": \"./" + config.buildPaths.html + "\"}}, \"Behat\\\\MinkExtension\": {\"base_url\": \"" + config.siteUrls[key] + "\", \"zombie\": {\"node_modules_path\": \"" + process.cwd() + "/node_modules/\"}}}}"
               })
             }, options)
           }