+++ tasks/operations.js
@@ -1,14 +1,14 @@
 module.exports = function(grunt) {
-
   /**
    * Define operations tasks as configured via scripts in Gruntconfig.json
    *
    * These are ad hoc scripts with minimal facilitation from Grunt.
    */
+  var _ = require('lodash');
 
   if (grunt.config('config.scripts')) {
     var scripts = grunt.config('config.scripts');
-    for (var key in scripts) {
+    _.each(scripts, function(script, key) {
       grunt.registerTask(key, 'Perform the configured "' + key + '" task.', function() {
         var task = require('../lib/scripts')(grunt)
           .handle(grunt.config('config'), this.name, 'ops');
@@ -22,7 +22,6 @@ module.exports = function(grunt) {
         task: key,
         group: 'Operations'
       });
-    }
+    });
   }
-
-}
+};