+++ test/test_assets_d8/src/modules/test.php
@@ -1,4 +1,7 @@
 <?php
+/**
+ * @file
+ * A doc comment for coding standards.
+ */
 
 $x = 10;
-